move lots of linux compatibility code into a dedicated file
